LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-buttons-containers-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<div id="container"></div>
<style type="text/css" media="screen">

body {
    margin: 0;
}

body > * {
    position: absolute;
}

#container {
    position: relative;
    top: 0;
    left: 0;
    width: 800px;
    height: 600px;
}

</style>
<script type="text/javascript">

description("Testing the <code>MacOSFullscreenMediaControls</code> computed styles.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSFullscreenMediaControls;
const centerContainer = mediaControls.controlsBar.children[2];
const rightContainer = mediaControls.controlsBar.children[3];
const buttons = centerContainer.buttons.concat(rightContainer.buttons);

document.getElementById("container").appendChild(mediaControls.element);

debug("");
shouldBecomeEqual("centerContainer.element.getBoundingClientRect().width", "178", () => {
    shouldBe("centerContainer.element.getBoundingClientRect().left", "311");
    shouldBe("centerContainer.element.getBoundingClientRect().top", "500");
    shouldBe("centerContainer.element.getBoundingClientRect().height", "44");

    debug("");
    shouldBecomeEqual("Math.floor(rightContainer.element.getBoundingClientRect().width)", "138", () => {
        shouldBeCloseTo("rightContainer.element.getBoundingClientRect().left", 496, 0.5);
        shouldBe("rightContainer.element.getBoundingClientRect().top", "500");
        shouldBe("rightContainer.element.getBoundingClientRect().height", "44");

        debug("");
        mediaControls.element.remove();
        finishJSTest();
    });
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
