LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-buttons-containers-styles-expected.txt

Testing the MacOSFullscreenMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS centerContainer.element.getBoundingClientRect().width became 178
PASS centerContainer.element.getBoundingClientRect().left is 311
PASS centerContainer.element.getBoundingClientRect().top is 500
PASS centerContainer.element.getBoundingClientRect().height is 44

PASS Math.floor(rightContainer.element.getBoundingClientRect().width) became 138
PASS rightContainer.element.getBoundingClientRect().left is within 0.5 of 496
PASS rightContainer.element.getBoundingClientRect().top is 500
PASS rightContainer.element.getBoundingClientRect().height is 44

PASS successfullyParsed is true

TEST COMPLETE
